requirements.txt
flask
requests
flask_sqlalchemy
flask_login
Font-Awesome-Flask
wtforms
flask_wtf
Flask-OAuthlib
Werkzeug>=2.2,<3.0
SQLAlchemy-serializer
flask_socketio
gunicorn
flask-cors
flask_httpauth
PyJWT==2.4.0
pytest
pytest_mock
Flask-Testing